% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_post_trends}
\alias{plot_post_trends}
\title{Visualize Posting Trends}
\usage{
plot_post_trends(data)
}
\arguments{
\item{data}{A data frame with social media activity data.}
}
\value{
A ggplot object showing the posting trends.
}
\description{
This function creates a line plot visualizing the number of posts over time.
}
\examples{
plot_post_trends(social_data)
}
