% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_data.R
\docType{data}
\name{social_data}
\alias{social_data}
\title{Social Media Engagement Data}
\format{
A data frame with 10 rows and 5 variables:
\describe{
  \item{date}{Date of the social media activity.}
  \item{posts}{Number of posts on the given date.}
  \item{likes}{Number of likes received on the given date.}
  \item{comments}{Number of comments received on the given date.}
  \item{shares}{Number of shares received on the given date.}
}
}
\source{
Synthetic data created for demonstration purposes.
}
\usage{
data(social_data)
}
\description{
A dataset containing synthetic data representing social media activity over a period.
It includes daily counts of posts, likes, comments, and shares.
}
\keyword{datasets}
